requirements.txt
numpy>=1.21.0
torch>=2.0.0
PyQt5>=5.15.0
Pillow>=8.0.0
cupy-cuda12x>=12.0.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
pandas>=1.3.0
optuna>=3.0.0
tensorboard>=2.6.0
tqdm>=4.62.0
scipy>=1.7.0 